refactor: merge hook-esm and hook-cjs together
It is simpler to handle the transformation of both module types together
in Jest.
