--- components/hook-cjs/void/.env
@@ -1,1 +1,0 @@
-browser
