--- components/hook-esm/node/.env
@@ -1,2 +1,0 @@
-test
-node
