--- components/hook-esm/node/.ordering
@@ -1,2 +1,0 @@
-stringify.mjs
-index.mjs
