+++ components/hook-module/node/.ordering
@@ -1,0 +1,4 @@
+stringify.mjs
+cjs.mjs
+esm.mjs
+index.mjs
