+++ components/hook-module/node/esm.mjs
@@ -1,14 +1,14 @@
 /* eslint-disable no-import-assign */
 
 // NB: To have any effect, this module needs `lib/node/loader-esm.mjs` to
-// executed as an `--experimented-loader`.
+// be ran as an `--experimented-loader`.
 // It could be worthwhile to detect if this is indeed the case.
 // We could use https://nodejs.org/api/modules.html#moduleispreloading.
 
 // This module introduces state to `lib/node/loader-esm.mjs`.
 // It is not very testable and hard to maintain.
-// It could be avoided by bubling this function instead.
-// However this would complexify the api.
+// It could be avoided by passing this function to the entry point instead.
+// However this would render the api more complex.
 // So let's keep it like for the moment.
 
 import {
@@ -38,11 +38,15 @@ export const hook = (agent, { hooks: { esm } }) => {
     hooked = esm;
     const transformModule = (url, format, content) =>
       format === "module"
-        ? instrument(agent, {
-            url,
-            type: "module",
-            content: stringifyContent(content),
-          })
+        ? instrument(
+            agent,
+            {
+              url,
+              type: "module",
+              content: stringifyContent(content),
+            },
+            null,
+          )
         : content;
     hooks.load = async (url, context, nextAsync) => {
       const { format, source } = await nextAsync(url, context, nextAsync);
