+++ components/hook-module/node/stringify.test.mjs
